% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cymruservices-package.r
\docType{package}
\name{cymruservices}
\alias{cymruservices}
\alias{cymruservices-package}
\title{cymruservices is an R package that provides interfaces to various
\href{http://www.team-cymru.org/services.html}{Team Cymru Services} including
The Bogon Refrerence, The IP to ASN Mapping Project and The Malware Hash Registry}
\description{
cymruservices is an R package that provides interfaces to various
\href{http://www.team-cymru.org/services.html}{Team Cymru Services} including
The Bogon Refrerence, The IP to ASN Mapping Project and The Malware Hash Registry
}
\note{
A direct connection to TCP Port 43 (WHOIS) is required for most of these
      API functions to work properly.
}
\author{
Bob Rudis (@hrbrmstr)
}

