% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zflush.r
\name{flush}
\alias{flush}
\title{Flush cached results}
\usage{
flush(..., quiet = TRUE)
}
\arguments{
\item{...}{strings naming cached results to flush. Can be any of
"\code{origin}", "\code{peer}", "\code{asn}", "\code{v4_bogons}",
"\code{v6_bogons}" or "\code{hash}". If no parameters
are specified all caches will be flushed.}

\item{quiet}{if \code{TRUE} no diagnostic or informative messages will be
displayed. If \code{FALSE} warnings for unknown cache names and
invalidation progress for valid caches will be displayed if
the session is interactive.}
}
\description{
Within a given R session, it will be highly unlikely that API responses
to calls to Team Cymru services will change if the parameters have
not varied (i.e. you use the same vector of IP addresses again). To
respect the resources that have beeen freely provided, all the API
functions cache their results.\cr
\cr
It may be advantageous or necessary to invalidate one or more of these
caches. This function allows for the invalidation of one or more (or all)
caches.
}
\note{
Invalid cache names will be ignored. If \code{quiet} is \code{FALSE}
      and \code{flush} was called from an interactive session invalid
      cache names will be noted.\cr\cr Also, you will still need to
      \code{force} the reloading of bogon lists if you are within the 4
      hour window even if you invalided the memoised cache.
}
\examples{
\dontrun{
flush("peer", "origin")
flush()
}
}
